Move Windows FS notifier code to separate file
