+++ src/main/cpp/win.cpp
@@ -18,9 +18,7 @@
 
 #include "native.h"
 #include "generic.h"
-#include <windows.h>
-#include <Shlwapi.h>
-#include <wchar.h>
+#include "win.h"
 
 #define ALL_COLORS (FOREGROUND_BLUE|FOREGROUND_RED|FOREGROUND_GREEN)
 
@@ -41,103 +39,7 @@ int map_error_code(int error_code) {
     return FAILURE_GENERIC;
 }
 
-jstring wchar_to_java(JNIEnv* env, const wchar_t* chars, size_t len, jobject result) {
-    if (sizeof(wchar_t) != 2) {
-        mark_failed_with_message(env, "unexpected size of wchar_t", result);
-        return NULL;
-    }
-    return env->NewString((jchar*)chars, len);
-}
 
-wchar_t* java_to_wchar(JNIEnv *env, jstring string, jobject result) {
-    jsize len = env->GetStringLength(string);
-    wchar_t* str = (wchar_t*)malloc(sizeof(wchar_t) * (len+1));
-    env->GetStringRegion(string, 0, len, (jchar*)str);
-    str[len] = L'\0';
-    return str;
-}
-
-//
-// Returns 'true' if the path of the form "X:\", where 'X' is a drive letter.
-//
-bool is_path_absolute_local(wchar_t* path, int path_len) {
-    if (path_len < 3) {
-        return false;
-    }
-    return (('a' <= path[0] && path[0] <= 'z') || ('A' <= path[0] && path[0] <= 'Z')) &&
-        path[1] == ':' &&
-        path[2] == '\\';
-}
-
-//
-// Returns 'true' if the path is of the form "\\server\share", i.e. is a UNC path.
-//
-bool is_path_absolute_unc(wchar_t* path, int path_len) {
-    if (path_len < 3) {
-        return false;
-    }
-    return path[0] == '\\' && path[1] == '\\';
-}
-
-//
-// Returns a UTF-16 string that is the concatenation of |prefix| and |path|.
-//
-wchar_t* add_prefix(wchar_t* path, int path_len, wchar_t* prefix) {
-    int prefix_len = wcslen(prefix);
-    int str_len = path_len + prefix_len;
-    wchar_t* str = (wchar_t*)malloc(sizeof(wchar_t) * (str_len + 1));
-    wcscpy_s(str, str_len + 1, prefix);
-    wcsncat_s(str, str_len + 1, path, path_len);
-    return str;
-}
-
-//
-// Returns a UTF-16 string that is the concatenation of |path| and |suffix|.
-//
-wchar_t* add_suffix(wchar_t* path, int path_len, wchar_t* suffix) {
-    int suffix_len = wcslen(suffix);
-    int str_len = path_len + suffix_len;
-    wchar_t* str = (wchar_t*)malloc(sizeof(wchar_t) * (str_len + 1));
-    wcsncpy_s(str, str_len + 1, path, path_len);
-    wcscat_s(str, str_len + 1, suffix);
-    return str;
-}
-
-//
-// Converts a Java string to a UNICODE path, including the Long Path prefix ("\\?\")
-// so that the resulting path supports paths longer than MAX_PATH (260 characters)
-//
-wchar_t* java_to_wchar_path(JNIEnv *env, jstring string, jobject result) {
-    // Copy the Java string into a UTF-16 string.
-    jsize len = env->GetStringLength(string);
-    wchar_t* str = (wchar_t*)malloc(sizeof(wchar_t) * (len+1));
-    env->GetStringRegion(string, 0, len, (jchar*)str);
-    str[len] = L'\0';
-
-    // Technically, this should be MAX_PATH (i.e. 260), except some Win32 API related
-    // to working with directory paths are actually limited to 240. It is just
-    // safer/simpler to cover both cases in one code path.
-    if (len <= 240) {
-        return str;
-    }
-
-    if (is_path_absolute_local(str, len)) {
-        // Format: C:\... -> \\?\C:\...
-        wchar_t* str2 = add_prefix(str, len, L"\\\\\?\\");
-        free(str);
-        return str2;
-    } else if (is_path_absolute_unc(str, len)) {
-        // In this case, we need to skip the first 2 characters:
-        // Format: \\server\share\... -> \\?\UNC\server\share\...
-        wchar_t* str2 = add_prefix(&str[2], len - 2, L"\\\\?\\UNC\\");
-        free(str);
-        return str2;
-    }
-    else {
-        // It is some sort of unknown format, don't mess with it
-        return str;
-    }
-}
 
 //
 // Returns 'true' if a file, given its attributes, is a Windows Symbolic Link.
@@ -1131,113 +1033,6 @@ Java_net_rubygrapefruit_platform_internal_jni_WindowsRegistryFunctions_getValueN
     RegCloseKey(key);
     free(subkeyStr);
     return true;
-}
-
-// File watching
-
-JavaVM* jvm = NULL;
-
-typedef struct watch_details {
-    HANDLE watchHandle;
-    HANDLE threadHandle;
-    jstring path;
-    jobject watcherCallback;
-} watch_details_t;
-
-DWORD WINAPI EventProcessingThread(LPVOID data) {
-    watch_details_t *details = (watch_details_t*) data;
-    while (TRUE) {
-        if (WaitForSingleObject(details->watchHandle, INFINITE) == WAIT_FAILED) {
-            // TODO Error handling
-            // mark_failed_with_errno(env, "could not wait for change notification", result);
-            break;
-        }
-        if (!FindNextChangeNotification(details->watchHandle)) {
-            if (GetLastError() == ERROR_INVALID_HANDLE) {
-                // Assumed closed
-                break;
-            }
-            // TODO Error handling
-            // mark_failed_with_errno(env, "could not schedule next change notification", result);
-            break;
-        }
-
-        JNIEnv* env;
-        int getEnvStat = jvm->GetEnv((void **)&env, JNI_VERSION_1_6);
-        if (getEnvStat == JNI_EDETACHED) {
-            int attachThreadStat = jvm->AttachCurrentThread((void **) &env, NULL);
-            if (attachThreadStat != JNI_OK) {
-                printf("~~~~ Problem with AttachCurrentThread: %d\n", attachThreadStat);
-                // TODO Error handling
-                // invalidStateDetected = true;
-                break;
-            }
-        } else if (getEnvStat == JNI_EVERSION) {
-            printf("~~~~ Problem with GetEnv: %d\n", getEnvStat);
-            // TODO Error handling
-            // invalidStateDetected = true;
-            break;
-        }
-
-        wchar_t* pathStr = java_to_wchar_path(env, details->path, NULL);
-        printf("~~~~ Changes: %ls\n", pathStr);
-        free(pathStr);
-
-        jclass callback_class = env->GetObjectClass(details->watcherCallback);
-        jmethodID methodCallback = env->GetMethodID(callback_class, "pathChanged", "(Ljava/lang/String;)V");
-        env->CallVoidMethod(details->watcherCallback, methodCallback, details->path);
-    }
-    return 0;
-}
-
-JNIEXPORT jobject JNICALL
-Java_net_rubygrapefruit_platform_internal_jni_WindowsFileEventFunctions_startWatch(JNIEnv *env, jclass target, jstring path, jobject javaCallback, jobject result) {
-    wchar_t* pathStr = java_to_wchar_path(env, path, result);
-    HANDLE watchHandle = FindFirstChangeNotificationW(
-        pathStr,
-        TRUE,
-        FILE_NOTIFY_CHANGE_FILE_NAME | FILE_NOTIFY_CHANGE_DIR_NAME | FILE_NOTIFY_CHANGE_ATTRIBUTES | FILE_NOTIFY_CHANGE_SIZE | FILE_NOTIFY_CHANGE_LAST_WRITE
-    );
-    // free(pathStr);
-    if (watchHandle == INVALID_HANDLE_VALUE) {
-        mark_failed_with_errno(env, "could not open change notification", result);
-        return NULL;
-    }
-
-    // TODO Should this be somewhere global?
-    int jvmStatus = env->GetJavaVM(&jvm);
-    if (jvmStatus < 0) {
-        mark_failed_with_errno(env, "Could not store jvm instance.", result);
-        return NULL;
-    }
-
-    watch_details_t* details = (watch_details_t*)malloc(sizeof(watch_details_t));
-    details->watchHandle = watchHandle;
-    details->path = (jstring) env->NewGlobalRef(path);
-    details->watcherCallback = env->NewGlobalRef(javaCallback);
-
-    details->threadHandle = CreateThread(
-        NULL,                   // default security attributes
-        0,                      // use default stack size
-        EventProcessingThread,  // thread function name
-        details,                // argument to thread function
-        0,                      // use default creation flags
-        NULL                    // the thread identifier
-    );
-
-    jclass clsWatch = env->FindClass("net/rubygrapefruit/platform/internal/jni/WindowsFileEventFunctions$WatchImpl");
-    jmethodID constructor = env->GetMethodID(clsWatch, "<init>", "(Ljava/lang/Object;)V");
-    return env->NewObject(clsWatch, constructor, env->NewDirectByteBuffer(details, sizeof(details)));
-}
-
-JNIEXPORT void JNICALL
-Java_net_rubygrapefruit_platform_internal_jni_WindowsFileEventFunctions_stopWatch(JNIEnv *env, jclass target, jobject detailsObj, jobject result) {
-    watch_details_t* details = (watch_details_t*)env->GetDirectBufferAddress(detailsObj);
-    env->DeleteGlobalRef(details->path);
-    env->DeleteGlobalRef(details->watcherCallback);
-    FindCloseChangeNotification(details->watchHandle);
-    CloseHandle(details->threadHandle);
-    free(details);
 }
 
 #endif
