+++ src/main/cpp/win_fsnotifier.cpp
@@ -3,12 +3,15 @@
 
 #ifdef _WIN32
 
+#include "native.h"
+#include "generic.h"
+#include "win.h"
+
 #include <ctype.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <wchar.h>
 #include <Windows.h>
 
 typedef struct {
@@ -391,6 +394,113 @@ int main(int argc, char *argv[]) {
     DeleteCriticalSection(&csOutput);
 
     return 0;
+}
+
+// File watching
+
+JavaVM* jvm = NULL;
+
+typedef struct watch_details {
+    HANDLE watchHandle;
+    HANDLE threadHandle;
+    jstring path;
+    jobject watcherCallback;
+} watch_details_t;
+
+DWORD WINAPI EventProcessingThread(LPVOID data) {
+    watch_details_t *details = (watch_details_t*) data;
+    while (TRUE) {
+        if (WaitForSingleObject(details->watchHandle, INFINITE) == WAIT_FAILED) {
+            // TODO Error handling
+            // mark_failed_with_errno(env, "could not wait for change notification", result);
+            break;
+        }
+        if (!FindNextChangeNotification(details->watchHandle)) {
+            if (GetLastError() == ERROR_INVALID_HANDLE) {
+                // Assumed closed
+                break;
+            }
+            // TODO Error handling
+            // mark_failed_with_errno(env, "could not schedule next change notification", result);
+            break;
+        }
+
+        JNIEnv* env;
+        int getEnvStat = jvm->GetEnv((void **)&env, JNI_VERSION_1_6);
+        if (getEnvStat == JNI_EDETACHED) {
+            int attachThreadStat = jvm->AttachCurrentThread((void **) &env, NULL);
+            if (attachThreadStat != JNI_OK) {
+                printf("~~~~ Problem with AttachCurrentThread: %d\n", attachThreadStat);
+                // TODO Error handling
+                // invalidStateDetected = true;
+                break;
+            }
+        } else if (getEnvStat == JNI_EVERSION) {
+            printf("~~~~ Problem with GetEnv: %d\n", getEnvStat);
+            // TODO Error handling
+            // invalidStateDetected = true;
+            break;
+        }
+
+        wchar_t* pathStr = java_to_wchar_path(env, details->path, NULL);
+        printf("~~~~ Changes: %ls\n", pathStr);
+        free(pathStr);
+
+        jclass callback_class = env->GetObjectClass(details->watcherCallback);
+        jmethodID methodCallback = env->GetMethodID(callback_class, "pathChanged", "(Ljava/lang/String;)V");
+        env->CallVoidMethod(details->watcherCallback, methodCallback, details->path);
+    }
+    return 0;
+}
+
+JNIEXPORT jobject JNICALL
+Java_net_rubygrapefruit_platform_internal_jni_WindowsFileEventFunctions_startWatch(JNIEnv *env, jclass target, jstring path, jobject javaCallback, jobject result) {
+    wchar_t* pathStr = java_to_wchar_path(env, path, result);
+    HANDLE watchHandle = FindFirstChangeNotificationW(
+        pathStr,
+        TRUE,
+        FILE_NOTIFY_CHANGE_FILE_NAME | FILE_NOTIFY_CHANGE_DIR_NAME | FILE_NOTIFY_CHANGE_ATTRIBUTES | FILE_NOTIFY_CHANGE_SIZE | FILE_NOTIFY_CHANGE_LAST_WRITE
+    );
+    // free(pathStr);
+    if (watchHandle == INVALID_HANDLE_VALUE) {
+        mark_failed_with_errno(env, "could not open change notification", result);
+        return NULL;
+    }
+
+    // TODO Should this be somewhere global?
+    int jvmStatus = env->GetJavaVM(&jvm);
+    if (jvmStatus < 0) {
+        mark_failed_with_errno(env, "Could not store jvm instance.", result);
+        return NULL;
+    }
+
+    watch_details_t* details = (watch_details_t*)malloc(sizeof(watch_details_t));
+    details->watchHandle = watchHandle;
+    details->path = (jstring) env->NewGlobalRef(path);
+    details->watcherCallback = env->NewGlobalRef(javaCallback);
+
+    details->threadHandle = CreateThread(
+        NULL,                   // default security attributes
+        0,                      // use default stack size
+        EventProcessingThread,  // thread function name
+        details,                // argument to thread function
+        0,                      // use default creation flags
+        NULL                    // the thread identifier
+    );
+
+    jclass clsWatch = env->FindClass("net/rubygrapefruit/platform/internal/jni/WindowsFileEventFunctions$WatchImpl");
+    jmethodID constructor = env->GetMethodID(clsWatch, "<init>", "(Ljava/lang/Object;)V");
+    return env->NewObject(clsWatch, constructor, env->NewDirectByteBuffer(details, sizeof(details)));
+}
+
+JNIEXPORT void JNICALL
+Java_net_rubygrapefruit_platform_internal_jni_WindowsFileEventFunctions_stopWatch(JNIEnv *env, jclass target, jobject detailsObj, jobject result) {
+    watch_details_t* details = (watch_details_t*)env->GetDirectBufferAddress(detailsObj);
+    env->DeleteGlobalRef(details->path);
+    env->DeleteGlobalRef(details->watcherCallback);
+    FindCloseChangeNotification(details->watchHandle);
+    CloseHandle(details->threadHandle);
+    free(details);
 }
 
 #endif
